chore (refs T22614): move assessmentTable Component to client/js.
